Updated GameState with real values from turns Added active_client_id to the UpdateGameStateNetMessage

+++ server/src/game/turns/TurnManager.h
@@ -29,8 +29,8 @@ public:
 
     size_t get_current_client_id() const;
     size_t get_current_worm_id() const;
-    size_t get_remaining_game_time() const;
-    size_t get_remaining_turn_time() const;
+    float get_remaining_game_time() const;
+    float get_remaining_turn_time() const;
 
     void add_player(size_t client_id, const std::list<size_t>& worm_ids_from_client);
     void remove_worm(size_t worm_id);
